refactor(pages/Create, Details, Edit ): Altera importacao e implementa token aleatorio
